Visual Updates
Several visual updates have been added to the game.

No code changes, but some new decorations, textures, and a skybox have all been added.

Lighting has also been tweaked to it casts a perfect shadow of the player which can be used to help aid the player in making jumps.

The design document file has been converted from Pages to Microsoft Word, so it may be opened on more devices a little easier.